# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_analyze_ticket_image_ai_vision_ticket_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_analyze_ticket_image_ai_vision_ticket_post",
        "type": "object"
      },
      "Body_analyze_ticket_image_vision_ticket_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_analyze_ticket_image_vision_ticket_post",
        "type": "object"
      },
      "Body_classify_seat_view_image_ai_vision_seat_view_classify_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_classify_seat_view_image_ai_vision_seat_view_classify_post",
        "type": "object"
      },
      "Body_classify_seat_view_image_vision_seat_view_classify_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_classify_seat_view_image_vision_seat_view_classify_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SeatViewClassification": {
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "SeatViewClassification",
        "type": "object"
      },
      "TicketInfo": {
        "properties": {
          "awayTeam": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awayteam"
          },
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "homeTeam": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hometeam"
          },
          "peopleCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peoplecount"
          },
          "price": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "reservationNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reservationnumber"
          },
          "row": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row"
          },
          "seat": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seat"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "stadium": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stadium"
          },
          "time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time"
          }
        },
        "title": "TicketInfo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "InternalApiKey": {
        "description": "AI 내부 호출용 키. Authorization Bearer 토큰을 사용할 수 있습니다.",
        "in": "header",
        "name": "X-Internal-Api-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "KBO AI Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/ai/vision/seat-view-classify": {
      "post": {
        "description": "Classify an uploaded baseball-related image for seat-view moderation.",
        "operationId": "classify_seat_view_image_ai_vision_seat_view_classify_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Internal-Api-Key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Api-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_classify_seat_view_image_ai_vision_seat_view_classify_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatViewClassification"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "InternalApiKey": []
          }
        ],
        "summary": "Classify Seat View Image",
        "tags": [
          "vision"
        ]
      }
    },
    "/ai/vision/ticket": {
      "post": {
        "description": "Analyzes an uploaded ticket image using Gemini Vision (Native or OpenRouter) to extract details.",
        "operationId": "analyze_ticket_image_ai_vision_ticket_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Internal-Api-Key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Api-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_ticket_image_ai_vision_ticket_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "InternalApiKey": []
          }
        ],
        "summary": "Analyze Ticket Image",
        "tags": [
          "vision"
        ]
      }
    },
...
    "/vision/seat-view-classify": {
      "post": {
        "deprecated": true,
        "description": "Classify an uploaded baseball-related image for seat-view moderation.",
        "operationId": "classify_seat_view_image_vision_seat_view_classify_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Internal-Api-Key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Api-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_classify_seat_view_image_vision_seat_view_classify_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatViewClassification"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "InternalApiKey": []
          }
        ],
        "summary": "Classify Seat View Image",
        "tags": [
          "vision"
        ]
      }
    },
    "/vision/ticket": {
      "post": {
        "deprecated": true,
        "description": "Analyzes an uploaded ticket image using Gemini Vision (Native or OpenRouter) to extract details.",
        "operationId": "analyze_ticket_image_vision_ticket_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Internal-Api-Key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Internal-Api-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_ticket_image_vision_ticket_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "InternalApiKey": []
          }
        ],
        "summary": "Analyze Ticket Image",
        "tags": [
          "vision"
        ]
      }
    }
  }
}